Reorganize project references and add visualization
The Test Data and Infrastructure projects flowed in the wrong direction because of where the Seed Data Helpers class was located.
