+++ src/Infrastructure/Infrastructure.csproj
@@ -18,7 +18,7 @@
 
     <ItemGroup>
         <ProjectReference Include="..\AppServices\AppServices.csproj" />
-        <ProjectReference Include="..\Domain\Domain.csproj" />
+        <ProjectReference Include="..\TestData\TestData.csproj" />
     </ItemGroup>
 
     <ItemGroup>
