+++ src/TestData/Properties/AssemblyInfo.cs
@@ -1,6 +1,7 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("AppServicesTests")]
+[assembly: InternalsVisibleTo("Infrastructure")]
 [assembly: InternalsVisibleTo("IntegrationTests")]
 [assembly: InternalsVisibleTo("LocalRepository")]
 [assembly: InternalsVisibleTo("LocalRepositoryTests")]
